Updated logging and fixed issues around editing the dom. Additionally added support so that this project will run out of box on BLC module projects
Logging wasn't very good around telling you which file broke and which xpath specifically broke so that was updated. Also logged not having a file path as an error since it's required
Before to edit the dom document.getFirstChild() was used. This breaks if something like BLC's license is at the top because it's counted as the first child. Now the "beans" node is explicitly found
Previously this project wouldn't run on BLC modules because their app contexts don't start with "applicationContext". Now it just requires the file to have the work applicationContext in it and it be an xml file to be considered for migration